+++ base/uk.ac.stfc.isis.ibex.scriptgenerator/src/uk/ac/stfc/isis/ibex/scriptgenerator/table/ActionsTable.java
@@ -224,10 +224,10 @@ public class ActionsTable extends ModelObject {
      * 
      * @param validityErrors The hashmap to set estimated time based on.
      */
-    public void setEstimatedTime(Map<Integer, Double> estimatedTimes) {
+    public void setEstimatedTime(Map<Integer, Integer> estimatedTimes) {
         for (int i = 0; i < actions.size(); i++) {
             if (estimatedTimes.containsKey(i)) {
-                Optional<Double> estimate = Optional.of(estimatedTimes.get(i));
+                Optional<Integer> estimate = Optional.of(estimatedTimes.get(i));
                 actions.get(i).setEstimatedTime(estimate);
             } else {
                 actions.get(i).setEstimatedTime(Optional.empty());
@@ -259,9 +259,9 @@ public class ActionsTable extends ModelObject {
      * @return An Optional containing the total if at least one action has been estimated,
      *         empty optional otherwise
      */
-    public Optional<Double> getTotalEstimatedTime() {
+    public Optional<Integer> getTotalEstimatedTime() {
         boolean isAnyActionEstimated = false;
-        Double total = 0.0;
+        Integer total = 0;
         
         for (ScriptGeneratorAction action : actions) {
             if (action.getEstimatedTime().isPresent()) {
